fix: 合并 PR #779/#787/#790/#792 评审 — SQL 通配符转义 + 整数溢出加固 + 去冗余 useMemo
整合 4 个历史 PR 及其代码评审建议为单一变更集：

- SQL 通配符注入（#792）：qa_repo.go ListDocuments 关键字用 dbutil.EscapeLike
  转义 % _ \ 并显式 ESCAPE '\'，杜绝通配符注入触发的全表扫描 / DoS。
- 整数溢出（#787 取代 #790，CWE-190 / G115）：container_monitor.go 将 Docker
  内存 Usage/Limit 截断到 math.MaxInt64 后再转 int64，MemoryPercent 改用截断值；
  system_monitor.go 的 Statfs Blocks/Bavail 统一上转 uint64 再相乘。
  采纳评审反馈：fetchContainers 全局汇总改用饱和累加，避免多个未限制容器
  的 MaxInt64 相加再次溢出为负。
- 去冗余 useMemo（#779）：useCachedImage 的 .trim() 与 Avatar 轻量校验
  去掉 useMemo 包装，省去 hook 开销。

未采纳 #792 评审「省略 ESCAPE '\'」建议：PG17 默认 standard_conforming_strings=on
下显式声明更自文档化且与 EscapeLike 的反斜杠转义字符严格对齐，予以保留。

📄 文档影响：[已更新 CHANGELOG.md]

+++ apps/server-go/internal/service/system_monitor.go
@@ -271,8 +271,9 @@ func (s *SystemMonitorService) collectDisk() DiskMetrics {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return DiskMetrics{Path: path}
 	}
-	total := stat.Blocks * uint64(stat.Bsize)
-	free := stat.Bavail * uint64(stat.Bsize)
+	// Blocks/Bavail 的类型随架构而异（部分平台为有符号），先统一上转 uint64 再相乘，避免溢出。
+	total := uint64(stat.Blocks) * uint64(stat.Bsize)
+	free := uint64(stat.Bavail) * uint64(stat.Bsize)
 	used := total - free
 	var pct float64
 	if total > 0 {
